issue #5747 refactor content scripts for Thunderbird (#5748)
* issue 5747 initial refactor

* removed dated comment

* cleanup

* added a generic folder

* remove unused notifyMurdered parameter

* added comment

* upload debug artifacts

* remove verbose google-mock listening logs
